refactor(oauth): separate captions into their own file
